refactor(dsp-meta): improve structure
