+++ dsp-meta/src/api/mod.rs
@@ -1,3 +1,4 @@
 pub mod app;
 pub(crate) mod convert;
 mod handler;
+mod model;
