+++ dsp-meta/src/api/model/mod.rs
@@ -1,0 +1,1 @@
+mod project_metadata_dto;
